+++ game/ato/flight.py
@@ -11,6 +11,7 @@ from dcs.planes import C_101CC, C_101EB, Su_33, FA_18C_hornet, C_130J_30
 
 from game.dcs.aircrafttype import AircraftType
 from game.theater import ControlPoint, MissionTarget
+from game.utils import max_optional_distance
 from pydcs_extensions.hercules.hercules import Hercules
 from .flightmembers import FlightMembers
 from .flightroster import FlightRoster
@@ -43,6 +44,7 @@ if TYPE_CHECKING:
     from .flightwaypoint import FlightWaypoint
     from .package import Package
     from .starttype import StartType
+    from game.utils import Distance
 
 F18_TGP_PYLON: int = 4
 
@@ -287,6 +289,12 @@ class Flight(
     def any_member_has_weapon_of_type(self, weapon_type: WeaponType) -> bool:
         return any(
             m.loadout.has_weapon_of_type(weapon_type) for m in self.iter_members()
+        )
+
+    def max_standoff_range(self) -> Optional[Distance]:
+        """The longest stand-off launch range across the flight's members, if any."""
+        return max_optional_distance(
+            m.loadout.max_standoff_range() for m in self.iter_members()
         )
 
     def __repr__(self) -> str: